Add showBooks module
