+++ index.js
@@ -2,10 +2,10 @@ import generateUniqueId from './node_modules/generate-unique-id/generateUniqueId
 import getTimeNow from './modules/timeNow.js';
 import toggleSectionVisibility from './modules/toggleSection.js';
 import getStoredBooks from './modules/getStoredBooks.js';
+import showBooks from './modules/showBooks.js';
 
 const myBooks = [];
 const addBtn = document.getElementById('add-btn');
-const listBooks = document.getElementById('list-books');
 const addBooksLink = document.getElementById('add-books-link');
 const listBooksLink = document.getElementById('list-books-link');
 const contactLink = document.getElementById('contact-link');
@@ -18,37 +18,6 @@ const currentTime = document.getElementById('current-time');
 //   myBooks.push(...myBooksUpdated);
 //   window.location.reload();
 // };
-
-const showBooks = () => {
-  listBooks.innerHTML = '';
-
-  myBooks.forEach((myBook) => {
-    const myBookDiv = document.createElement('div');
-    myBookDiv.classList.add('row');
-
-    const bookDetails = document.createElement('h4');
-    bookDetails.textContent = `${myBook.bookTitle} by ${myBook.bookAuthor}`;
-
-    const removeBtn = document.createElement('button');
-    removeBtn.classList.add('remove-btn');
-    removeBtn.textContent = 'Remove';
-
-    // inline function for the event listener
-    removeBtn.addEventListener('click', () => {
-      const bookIDToRemove = myBook.bookID;
-      const myBooksUpdated = myBooks.filter((myBook) => myBook.bookID !== bookIDToRemove);
-      localStorage.setItem('myBooks', JSON.stringify(myBooksUpdated));
-      myBooks.length = 0;
-      myBooks.push(...myBooksUpdated);
-      // Call the showBooks function to update the displayed books
-      showBooks();
-    });
-
-    myBookDiv.appendChild(bookDetails);
-    myBookDiv.appendChild(removeBtn);
-    listBooks.appendChild(myBookDiv);
-  });
-};
 
 const addBook = () => {
   const bookID = generateUniqueId();
@@ -65,7 +34,7 @@ const addBook = () => {
   myBooks.push({ bookID, bookTitle, bookAuthor });
   localStorage.setItem('myBooks', JSON.stringify(myBooks));
 
-  showBooks();
+  showBooks(myBooks);
 
   // Clear the input fields values
   document.getElementById('title').value = '';
@@ -90,4 +59,4 @@ contactLink.addEventListener('click', () => {
 
 // Initial display of books
 getStoredBooks(myBooks);
-showBooks();
+showBooks(myBooks);
